# Encode optimization