--- a/Cargo.toml
+++ b/Cargo.toml
@@ -14,3 +14,7 @@ byteorder = "1.0.0"
 
 [dev-dependencies]
 rand = "=0.3.15"
+
+[profile.release]
+# Uncomment when using `perf record`
+#debug = true