--- a/examples/make_tables.rs
+++ b/examples/make_tables.rs
@@ -15,6 +15,7 @@ fn main() {
         // /
         .chain(0x2F..0x30)
         .collect();
+    print_encode_table(&standard_alphabet, "STANDARD_ENCODE", 0);
     print_decode_table(&standard_alphabet, "STANDARD_DECODE", 0);
 
     // A-Z
@@ -28,8 +29,20 @@ fn main() {
         // _s
         .chain(0x5F..0x60)
         .collect();
+    print_encode_table(&url_alphabet, "URL_SAFE_ENCODE", 0);
     print_decode_table(&url_alphabet, "URL_SAFE_DECODE", 0);
 
+}
+
+fn print_encode_table(alphabet: &[u8], const_name: &str, indent_depth: usize) {
+    println!("{:width$}pub const {}: &'static [u8; 64] = &[", "", const_name, width=indent_depth);
+
+    for (i, b) in alphabet.iter().enumerate() {
+        println!("{:width$}{}, // input {} (0x{:X}) => '{}' (0x{:X})", "",
+                 b, i, i, String::from_utf8(vec!(*b as u8)).unwrap(), b, width=indent_depth + 4);
+    }
+
+    println!("{:width$}];", "", width=indent_depth);
 }
 
 fn print_decode_table(alphabet: &[u8], const_name: &str, indent_depth: usize) {
@@ -48,7 +61,7 @@ fn print_decode_table(alphabet: &[u8], const_name: &str, indent_depth: usize) {
             None => ("INVALID_VALUE".to_string(),
                     format!("input {} (0x{:X})", ascii_byte, ascii_byte)),
             Some(v) => (format!("{}", *v),
-                        format!("input {} (0x{:X} char '{}') = {} (0x{:X})",
+                        format!("input {} (0x{:X} char '{}') => {} (0x{:X})",
                                 ascii_byte,
                                 ascii_byte,
                                 String::from_utf8(vec!(ascii_byte as u8)).unwrap(), *v, *v))
@@ -56,5 +69,5 @@ fn print_decode_table(alphabet: &[u8], const_name: &str, indent_depth: usize) {
 
         println!("{:width$}{}, // {}", "", value, comment, width=indent_depth + 4);
     }
-    println!("];")
+    println!("{:width$}];", "", width=indent_depth);
 }