The good news is that it's a lot faster. The bad news is that it took a fair amount of unsafe, so double check that my pointer-wrangling isn't going to ruin the world.

Before:

```
test encode_100b            ... bench:         218 ns/iter (+/- 9) = 458 MB/s
test encode_100b_reuse_buf  ... bench:         280 ns/iter (+/- 80) = 357 MB/s
test encode_10mib           ... bench:  24,247,195 ns/iter (+/- 138,068) = 432 MB/s
test encode_10mib_reuse_buf ... bench:  21,752,566 ns/iter (+/- 33,576) = 482 MB/s
test encode_30mib           ... bench:  72,667,077 ns/iter (+/- 147,145) = 432 MB/s
test encode_30mib_reuse_buf ... bench:  65,158,430 ns/iter (+/- 38,008) = 482 MB/s
test encode_3b              ... bench:          24 ns/iter (+/- 0) = 125 MB/s
test encode_3b_reuse_buf    ... bench:           9 ns/iter (+/- 0) = 333 MB/s
test encode_3kib            ... bench:       6,193 ns/iter (+/- 6) = 496 MB/s
test encode_3kib_reuse_buf  ... bench:       6,177 ns/iter (+/- 6) = 497 MB/s
test encode_3mib            ... bench:   7,096,823 ns/iter (+/- 105,388) = 443 MB/s
test encode_3mib_reuse_buf  ... bench:   6,415,920 ns/iter (+/- 529,768) = 490 MB/s
test encode_500b            ... bench:       1,018 ns/iter (+/- 1) = 491 MB/s
test encode_500b_reuse_buf  ... bench:       1,002 ns/iter (+/- 1) = 499 MB/s
test encode_50b             ... bench:         115 ns/iter (+/- 0) = 434 MB/s
test encode_50b_reuse_buf   ... bench:          98 ns/iter (+/- 0) = 510 MB/s

 Performance counter stats for 'target/release/deps/benchmarks-a5a69865bf1a5765 --bench encode_10mib_reuse':

       9046.672897      task-clock (msec)         #    1.000 CPUs utilized          
                33      context-switches          #    0.004 K/sec                  
                 1      cpu-migrations            #    0.000 K/sec                  
           603,532      page-faults               #    0.067 M/sec                  
    34,356,962,385      cycles                    #    3.798 GHz                    
   <not supported>      stalled-cycles-frontend  
   <not supported>      stalled-cycles-backend   
    92,035,207,568      instructions              #    2.68  insns per cycle        
    10,790,184,917      branches                  # 1192.724 M/sec                  
         2,145,071      branch-misses             #    0.02% of all branches        

       9.046954261 seconds time elapsed
```

After:

```
test encode_100b            ... bench:          84 ns/iter (+/- 10) = 1190 MB/s
test encode_100b_reuse_buf  ... bench:          68 ns/iter (+/- 3) = 1470 MB/s
test encode_10mib           ... bench:   8,871,210 ns/iter (+/- 144,285) = 1181 MB/s
test encode_10mib_reuse_buf ... bench:   6,291,679 ns/iter (+/- 45,855) = 1666 MB/s
test encode_30mib           ... bench:  26,409,465 ns/iter (+/- 184,162) = 1191 MB/s
test encode_30mib_reuse_buf ... bench:  18,844,358 ns/iter (+/- 102,027) = 1669 MB/s
test encode_3b              ... bench:          26 ns/iter (+/- 1) = 115 MB/s
test encode_3b_reuse_buf    ... bench:          10 ns/iter (+/- 0) = 300 MB/s
test encode_3kib            ... bench:       1,709 ns/iter (+/- 48) = 1797 MB/s
test encode_3kib_reuse_buf  ... bench:       1,697 ns/iter (+/- 31) = 1810 MB/s
test encode_3mib            ... bench:   2,442,830 ns/iter (+/- 50,864) = 1287 MB/s
test encode_3mib_reuse_buf  ... bench:   1,737,254 ns/iter (+/- 40,098) = 1810 MB/s
test encode_500b            ... bench:         306 ns/iter (+/- 20) = 1633 MB/s
test encode_500b_reuse_buf  ... bench:         289 ns/iter (+/- 30) = 1730 MB/s
test encode_50b             ... bench:          53 ns/iter (+/- 51) = 943 MB/s
test encode_50b_reuse_buf   ... bench:          37 ns/iter (+/- 1) = 1351 MB/s

 Performance counter stats for 'target/release/deps/benchmarks-a5a69865bf1a5765 --bench encode_10mib_reuse':

       3998.986797      task-clock (msec)         #    1.000 CPUs utilized          
                25      context-switches          #    0.006 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
           603,536      page-faults               #    0.151 M/sec                  
    15,187,022,402      cycles                    #    3.798 GHz                    
   <not supported>      stalled-cycles-frontend  
   <not supported>      stalled-cycles-backend   
    50,464,295,566      instructions              #    3.32  insns per cycle        
     3,948,319,703      branches                  #  987.330 M/sec                  
           558,902      branch-misses             #    0.01% of all branches        

       3.999303452 seconds time elapsed
```

That's as fast as aklomp's C compiled with GCC, and faster than the C compiled with clang! See https://gist.github.com/marshallpierce/20351d44b29e238fecb821dc87251e5f for some notes with that data from my decode optimization work.